Added Observer design pattern using the human structure from the pelvis down to ankle
created my own Observable interface as the Observable class forces the use of inheritance instead of programming to an interface. it also doesn't allow instances to be created of the Observable class due to protecting methods so you have to subclass it. 

Am considering turning Ankle class into a Subject as this could be useful to notify the rest of the leg when the ankle touches the ground again, which would be when the step has been completed.